arbitrage_bot/requirements.txt
httpx==0.27.2
loguru==0.7.2
python-dotenv==1.0.1
thefuzz[speedup]==0.22.1
websockets==12.0
py_clob_client==0.25.0
